Updates from recent changes in master
