+++ src/js/components/ServicePlanProgressBar.js
@@ -13,8 +13,8 @@ class ServicePlanProgressBar extends React.Component {
     let activePhase = phases.getActiveIndex() + 1;
 
     return (
-      <span className="text-overflow-break-word">
-        {`Phase ${activePhase} of ${phases.getItems().length}: ${phases.getActive().getName()}`}
+      <span>
+        {`Phase ${activePhase} of ${phases.getItems().length}: ${phases.getActive().getName()} `}
         <a className="clickable" onClick={this.props.onViewDetailsClick}>
           View Details
         </a>
@@ -33,6 +33,7 @@ class ServicePlanProgressBar extends React.Component {
       fillClassName,
       fillWrapperClassName,
       phaseLabelClassName,
+      phaseProgressClassName,
       progress,
       servicePlan,
       statusLabelClassName
@@ -44,8 +45,12 @@ class ServicePlanProgressBar extends React.Component {
       fillWrapperClassName
     );
     let phaseLabelClasses = classNames(
-      'progress-bar-label-action',
+      'progress-bar-phase-details',
       phaseLabelClassName
+    );
+    let phaseProgressClasses = classNames(
+      'progress-bar-phase-progress',
+      phaseProgressClassName
     );
     let serviceStatusWrapperClasses = classNames({
       'is-erroneous': servicePlan.hasError(),
@@ -65,11 +70,13 @@ class ServicePlanProgressBar extends React.Component {
 
     return (
       <div className={wrapperClasses}>
-        <span className={statusLabelClasses}>
-          {this.getStatusLabel(servicePlan)}
-        </span>
-        <div className={fillWrapperClasses}>
-          <div className={fillClasses} style={{width: `${progress}%`}} />
+        <div className={phaseProgressClasses}>
+          <span className={statusLabelClasses}>
+            {this.getStatusLabel(servicePlan)}
+          </span>
+          <div className={fillWrapperClasses}>
+            <div className={fillClasses} style={{width: `${progress}%`}} />
+          </div>
         </div>
         <span className={phaseLabelClasses}>
           {this.getPhaseLabel(servicePlan)}
@@ -99,6 +106,10 @@ ServicePlanProgressBar.propTypes = {
   ]),
   onViewDetailsClick: React.PropTypes.func,
   phaseLabelClassName: React.PropTypes.oneOfType([
+    React.PropTypes.object,
+    React.PropTypes.string
+  ]),
+  phaseProgressClassName: React.PropTypes.oneOfType([
     React.PropTypes.object,
     React.PropTypes.string
   ]),
