+++ src/styles/index.less
@@ -151,6 +151,7 @@ Components
 @import "components/progressbar.less";
 @import "components/scrollbar.less";
 @import "components/service-plan-progress-modal.less";
+@import "components/service-plan-progress-bar.less";
 @import "components/service-sidebar.less";
 @import "components/side-panel-detail.less";
 @import "components/side-list.less";
