+++ tests/_fixtures/service-plan/2-phases-in-progress.json
@@ -1,5 +1,7 @@
 {
-  "errors": [],
+  "errors": [
+    
+  ],
   "phases": [
     {
       "blocks": [
